% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add1_models}
\alias{add1_models}
\title{Constructs a list of fits by adding predictors sequentially}
\usage{
add1_models(model, preds, data = NULL)
}
\arguments{
\item{model}{A linear model}

\item{preds}{Predictors to be added sequentially}

\item{data}{The dataset (optional)}
}
\value{
A list of linear fits
}
\description{
Constructs a list of fits by adding predictors sequentially
}
