% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createServer.R
\name{createERServer}
\alias{createERServer}
\title{A function which returns a shiny server for Exploratory Regression}
\usage{
createERServer(ERfit, ERdata = NULL, ERbarcols = RColorBrewer::brewer.pal(4,
  "Set2"), ERnpcpCols = 4, pvalOrder = F)
}
\arguments{
\item{ERfit}{the lm fit to be explored}

\item{ERdata}{the data used to fit the model. If NULL, attempts to extract from ERfit.}

\item{ERbarcols}{a vector of colours, one per term in lm.
Will be expanded via colorRampPalette if not the correct length.}

\item{ERnpcpCols}{number of colours for the PCP}

\item{pvalOrder}{if TRUE, re-arranges predictors in order of p-value}
}
\value{
a function
}
\description{
A function which returns a shiny server for Exploratory Regression
}
