% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUI.R
\name{createERUI}
\alias{createERUI}
\title{Constructs UI for Exploratory Regression app}
\usage{
createERUI(tablesOnly = F)
}
\arguments{
\item{tablesOnly}{if TRUE, shows Plots 1-3 only.}
}
\value{
the UI
}
\description{
Constructs UI for Exploratory Regression app
}
