% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{drop1_models}
\alias{drop1_models}
\title{Constructs a list of fits by dropping predictors from the supplied model}
\usage{
drop1_models(model, preds, data = NULL)
}
\arguments{
\item{model}{A linear model}

\item{preds}{Predictors to be dropped}

\item{data}{The dataset (optional)}
}
\value{
A list of linear fits
}
\description{
Constructs a list of fits by dropping predictors from the supplied model
}
