% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appPlots.R
\name{plotSeqSS}
\alias{plotSeqSS}
\title{Plots barcharts of sequential sums of squares of lm}
\usage{
plotSeqSS(fits, barcols = NULL, legend = F)
}
\arguments{
\item{fits}{list of lm objects}

\item{barcols}{a vector of colours, one per term in lms}

\item{legend}{TRUE or FALSE}
}
\value{
a ggplot
}
\description{
Plots barcharts of sequential sums of squares of lm
}
\examples{
plotSeqSS(list(fit1= lm(mpg ~ wt+hp+disp, data=mtcars),
fit2=lm(mpg ~ wt*hp*disp, data=mtcars)))
}
