% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appPlots.R
\name{termColours}
\alias{termColours}
\title{Constructs colour vector for model terms}
\usage{
termColours(f, pal = RColorBrewer::brewer.pal(4, "Set2"))
}
\arguments{
\item{f}{a model fit with term labels}

\item{pal}{use this palette}
}
\value{
a vector of colours. Residuals are given a grey color
}
\description{
Constructs colour vector for model terms
}
\examples{
termColours(lm(mpg ~ wt+hp, data=mtcars))
}
